Move creation of RPC clients to the RPC scehma class
